requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==7.1
djangorestframework==3.11
python-decouple==3.4
drf-yasg==1.20.0
django-rating
django-extensions
